+++ src/cli/reconcile.rs
@@ -1,10 +1,12 @@
+use rust_decimal::Decimal;
+
 use crate::db::get_connection;
 use crate::error::Result;
 use crate::fmt::money;
 use crate::reconciler;
 use crate::settings::get_data_dir;
 
-pub fn run(account: &str, month: &str, balance: f64) -> Result<()> {
+pub fn run(account: &str, month: &str, balance: Decimal) -> Result<()> {
     let conn = get_connection(&get_data_dir().join("nigel.db"))?;
     let result = reconciler::reconcile(&conn, account, month, balance)?;
